backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
youtube-transcript-api
openai>=1.0.0
